requirements.txt
Flask==2.3.3
numpy==1.26.4
google-generativeai==0.5.4
python-dotenv==1.0.1
google-cloud-storage==2.16.0
Flask-Cors==4.0.1
faiss-cpu==1.7.4
google-cloud-aiplatform==1.44.0
